3Dボックスの表示できるように対応

+++ GLLib.cpp
@@ -24,13 +24,12 @@ void LibraryEnd()
 	System::GetInstance()->Finalize();
 }
 
-int DrawBox(int posX, int posY, int width, int height, unsigned int color, int fillFlag)
+int DrawBox(float posX, float posY, float width, float height, unsigned int color, int fillFlag)
 {
 	return System::GetInstance()->DrawBox(posX, posY, width, height, color, fillFlag);
 }
 
-
-int DrawBox3D(int posX, int posY, int width, int height, unsigned int color, int fillFlag)
+int DrawBox3D(float posX, float posY, float width, float height, unsigned int color, int fillFlag)
 {
 	return System::GetInstance()->DrawBox3D(posX, posY, width, height, color, fillFlag);
 }
