+++ GLLib.h
@@ -9,6 +9,6 @@ bool ProcessMessage();
 void SwapBuffers();
 void LibraryEnd();
 
-int DrawBox(int, int, int, int, unsigned int, int);
-int DrawBox3D(int, int, int, int, unsigned int, int);
+int DrawBox(float, float, float, float, unsigned int, int);
+int DrawBox3D(float, float, float, float, unsigned int, int);
 }
