+++ System.h
@@ -22,8 +22,8 @@ public:
 	void Finalize();
 
 	// }Œ`•`‰æŠÖ˜A
-	int DrawBox(int, int, int, int, unsigned int, int);
-	int DrawBox3D(int, int, int, int, unsigned int, int);
+	int DrawBox(float, float, float, float, unsigned int, int);
+	int DrawBox3D(float, float, float, float, unsigned int, int);
 
 private:
 	System();
